+++ code/Input/inputSpecial.js
@@ -85,7 +85,7 @@ function inputSpecial() {
 			miniTilesDown = 0;
 			//clear lines
 			(function() {
-				for(y = 0; y < 20; y++){
+				for(y = 0; y < matrix.length; y++){
 					var foo = 0;
 					for (x = 0;x < 10; x++) {
 						if (matrix[y][x] > 0) {