+++ code/colCheck.js
@@ -2,7 +2,7 @@ function colCheck(newPosY, newPosX, newRotation) {
 	for (i=0; i < 4; i++) {
 		var x = minoData[game.currentPiece]["rotation"+newRotation][i][0]+newPosX;
 		var y = minoData[game.currentPiece]["rotation"+newRotation][i][1]+newPosY;
-		if (x < 0 || x > 9 || y > 19) {
+		if (x < 0 || x > 9 || y >= matrix.length) {
 			return true; //bounduary collision detected
 		} else {
 			if (matrix[y] !== undefined) {